backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==3.0.1

# Translation and NLP
deep-translator==1.11.4
textblob==0.17.1
gtts==2.5.0

# Document Processing
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
pdf2docx==0.5.6
python-docx==1.1.0
docx2pdf==0.1.8
img2pdf==0.5.1

# PDF Generation
reportlab==4.0.7

# Optional: Advanced features
# openai==1.3.5  # For GPT integration
# python-dotenv==1.0.0  # For environment variables
